requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
qutip>=4.7.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Data handling
h5py>=3.8.0
pyyaml>=6.0

# Real-time data and API
requests>=2.31.0
websockets>=11.0
aiohttp>=3.8.0

# Hardware interface (for prototype detector)
pyserial>=3.5

# LIGO/Virgo data access
gwpy>=3.0.0
ligo-gracedb>=2.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
